feats: streamlined the routing and fetching of tracks.
